Fixes RubyForge bug #9368, Illegal chars in about.yml
Bump version to 1.0.4

This fixes a bug where the ':' character in the about.yml description
was causing the parsing of about.yml to fail.

There was also a bad apostrophe char which was being escaped when the
yaml file was parsed.

You can now confirm that the about.yml parses in an 'irb' console with:

require 'yaml'
YAML.load(File.read('about.yml'))